Models with non-standard keys can't use the default whitelist.
The Session model, for example, extends AppModel and has a varchar
primary key. That means that the key value already exists and has to get
saved when the record is created, but the whitelist prevents an id value
from being saved via mass assignment. This fix creates no default
whitelist if the model being constructed has such a non-standard id.